reduction/combiner tests added, against logistic regression.

+++ applications/ml/lr/app.h
@@ -54,7 +54,7 @@ class LogisticRegression : public Application {
     LogisticRegression(const size_t& dimension_,
                        const size_t& iteration_num,
                        const size_t& partition_num,
-                       const size_t& sample_num_m,
+                       const double& sample_num_m,
                        const size_t& reduction_partition_num);
     ~LogisticRegression();
     virtual void Load();
@@ -62,7 +62,7 @@ class LogisticRegression : public Application {
     virtual size_t dimension();
     virtual size_t iteration_num();
     virtual size_t partition_num();
-    virtual size_t sample_num_m();
+    virtual double sample_num_m();
     virtual size_t reduction_partition_num();
     virtual size_t sample_num_per_partition();
     virtual bool automatic_reduction_active();
@@ -75,7 +75,7 @@ class LogisticRegression : public Application {
     size_t dimension_;
     size_t iteration_num_;
     size_t partition_num_;
-    size_t sample_num_m_;
+    double sample_num_m_;
     size_t reduction_partition_num_;
     size_t sample_num_per_partition_;
     bool automatic_reduction_active_;
